Implement dark mode/night mode
 * Add option to toggle dark mode/night mode for entire site
 * Saved on future visits
 * Add <noscript> message
 * Update some text

I wanted there to be a transition between the themes, and there was for
a little while, but apparently Chrome just sucks at doing transitions so
I had to scrap it for simplicity. It's not that big of a deal, I'm just
surprised that Chrome behaved so terrible while IE 11 worked just fine.